Add Undo Framework QML API
This commit introduces an initial implementation of a QML API for the
Undo Framework.

The QML plugin for this API exports three new QML elements: UndoStack,
UndoCommand and UndoPropertyCommand. The API for both the first and
the second are very similar to the QUndoStack and QUndoCommand C++
APIs. The thrid use the declarative nature of the QML elements to
undo/redo based on the values of a group of properties specified by
the user.

This commit also adds an example in 'examples/undo/quickundo' to show
how this API could be used.

Other patchs are expected soon to improve this API and also add more
examples/docs.

Change-Id: I9d290f514b360c721116c3937bd0e03d5f0eacbb
Reviewed-by: Anselmo L. S. Melo <anselmo.melo@openbossa.org>