Completion model QML API bootstrap
This patch introduces a QML-friendy API to the UiCompletionModel.

For now it is accepting just JS lists as models. Each element in the
JS list is converted to string and then passed to the completion model.
The plan is to make this element accept also other kind of models such
as ListModel, XMLListModel and C++ models in general.

There is an example in 'examples/models/quickcompletionmodel'.

Change-Id: Ia2c77fe778b6356cacf80bac78fabb2edd350f6e
Reviewed-by: Anselmo L. S. Melo <anselmo.melo@openbossa.org>